+++ projects/vedana/CHANGELOG.md
@@ -1,3 +1,11 @@
+# WIP
+
+* (vedana-etl) Implement parsing links from Grist links
+* (vedana-etl) Add datamodel node to graph
+* (vedana-gradio) Add GPT-5 to model selection
+* (vedana-gradio) Fix chat completion for gpt-5
+* (vedana-gradio) Remove obsolette UI components
+
 # 2025.08.07-dev.15
 
 * (vedana-core) Ported work with Memgraph to async